add parse_INFO function
